Ensure loss target has the correct type

+++ mini_nlp_framework/losses.py
@@ -7,12 +7,13 @@ LossFunction = Callable[[torch.Tensor, torch.Tensor], torch.Tensor]
 
 
 def flat_cross_entropy_loss(preds, target, ignore_index=-100):
-    #return F.cross_entropy(preds.view(-1, len(vocab.idx_to_word)), target.view(-1),
+    target = target.long()
     return F.cross_entropy(preds.view(-1, preds.shape[-1]), target.view(-1),
                            ignore_index=ignore_index)
 
 
 def flat_binary_cross_entropy_loss(preds, target):
+    target = target.float()
     return F.binary_cross_entropy_with_logits(preds.view(-1), target.view(-1))
